Make sure that the wheel build task fails if the wheel build fails
Due to the use of the pipe in the shell command the return code is
always zero even when the wheel build fails. This results in some
very obscure failures down the line which do not make it obvious
that the failure to build the appropriate wheel was the root cause.

In this patch we ensure that the shell will fail if the pip command
fails by setting the bash pipefail option.

Change-Id: Id539515557b439824e39ea2f693f62b913d5bdda
(cherry picked from commit 22914931395ac05cd03fcd9f92c7467061578927)