Override the pip index used for the wheel build
When doing the wheel build we need to ensure that we make
use of the pypi mirror directly.

When upgrading, if we leave it to the pip.conf file contents,
it will only use the contents of the pypiserver which locks
it down to the wheels built for the version we're upgrading
from. The wheel build therefore fails.

Leaving the pip.conf config in place is still good for
ensuring that anything installed on the repo server later uses
the pypiserver contents.

Change-Id: I5d1f33832c9738683b49384ea5ebfb571b526a81
(cherry picked from commit a25d1b41db49377c09b5e2c07ce3a38dd6e9df1c)