Allow repo host packages to be upgraded before repo build
To cater for a situation where the repo build packages are changing,
but the repo does not yet have the package built, we need to ensure
that the install task for it can fetch from pypi (or the designated
pypi mirror). To do this we try the local repo first, then fall back
to using the designated default index which defaults to pypi.

Change-Id: Iba7e9939250b24dc64fdd1883f963d0253ddad7e
(cherry picked from commit de46da9d144156d486367218204f8a2be1a31381)