tasks: repo_clone_git: Retry git-clone on failures
The async method of running the git-clone commands has the unfortunate
side effect that git-clones are not being retried in case of failures.
This breaks deployment on hosts that either have limited network
resources or when the remote server implements some sort of rate
limiting when they are being hit by many connections from the same host.

As such, lets get rid of the async method and use the regular
retry/until method to try and recover from the various git-clone
failures.

Change-Id: Iff0c3c589f5e0a935e0e01cd820083b76d43cc92