Update repo_build test-vars data
This updates the repo_build data to the current
package list output from the py_pkgs lookup.

It also changes the implementation of how the
test data is stored and consumed to make it easier
to maintain in the future.

Some line spacing is also added to the test playbook
to make the tasks easier to distinguish.

Depends-On: https://review.openstack.org/568621
Change-Id: Icf102a78fd0cf9aae3c09378235f2d91f7db8cdd