Update repo_build test-vars data
This updates the repo_build data to the current
package list output from the py_pkgs lookup.

It also changes the implementation of how the
test data is stored and consumed to make it easier
to maintain in the future.

Some line spacing is also added to the test playbook
to make the tasks easier to distinguish.

Conflicts:
1. tests/test-vars.yml
   due to the old test data removed being different
2. tests/test-data.yml
   due to the test data for stable/queens being different

Depends-On: https://review.openstack.org/568621
Change-Id: Icf102a78fd0cf9aae3c09378235f2d91f7db8cdd
(cherry picked from commit 98fcbb28ebd6e6f0cd21b2c43f556ffe74a00e23)