tasks: repo_build_wheels: Download get-pip.py if it does not exist
/opt/get-pip.py may not exist if the host uses the distribution
python-pip package. As such, we need to check if the file exists
and download it if it's missing before trying to copy it to the
repo build directory.

This is a combined backport of:
- https://review.openstack.org/570601
- https://review.openstack.org/572422
- https://review.openstack.org/580409
- https://review.openstack.org/583331

Related-Review: https://review.openstack.org/583925
Related-Bug: #1779534
Related-Bug: #1782051
Change-Id: I693b930bf337fd08931d835d57ffe9ecae904517
(cherry picked from commit ea4f8606df8a19edf5aa958a866f77ccafa62e47)