redhat: Add the EPEL repository for moreutils packages
The moreutils package is needed from EPEL so we install it inside
the role. Moreover, in order to minimize the potential EPEL vs RDO
package conflicts, we only allow a small subset of packages from the
EPEL repository. Finally, we include the nodepool playbook from the
tests repository in order to populate the necessary CI facts.

Also, we don't need the LXC development headers so drop them.

Change-Id: I287a7209d7840283e374a8bb14d9caa4f714baf0