repo_build: take distro version into account
When introducing a new repo_container with a different distro, using
the same directory for symlinks can result in weird builds. We should
prevent the links directory to be copied all willy-nilly, as suggested
here;

https://review.opendev.org/#/c/712784/
https://review.opendev.org/#/c/712787/

But we can further prevent bad builds by separating under distro-arch
under links with this patch.

Depends-On: Ieca6ec9bf898432c76c019a1d2d97c280cd85dce
Change-Id: I4ae6737f1ded4cbb680a380c8c309a0f5bf64723