repo_build: determine if there was a change when syncing files
If we sync a couple of item.paths over to the repo_all[0]. There is
definitely a change happening, so setting changed_when to false is
not correct.

During the with_items loop you can check each individual task under
the "register: _variable". After the task has completed with all
items, they're bundled up into _variable.results for future
processing.

rsync reports changes to stdout, and will remain empty if there are
none. So we can just check for stdout.

Depends-On: Ieca6ec9bf898432c76c019a1d2d97c280cd85dce
Change-Id: I2cf628d322e09ab7a0dc12557b07b1ab8d174b99